// src/public-api/model/addToCartDto.ts
/**
 * MISCore Shop API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AddToCartDto { 
    productId: number;
    quantity?: number;
    selectedOptions?: string | null;
}