// src/public-api/model/productDto.ts
/**
 * MISCore Shop API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProductTagDto } from './productTagDto';
import { ProductSpecificationDto } from './productSpecificationDto';


export interface ProductDto { 
    id?: number;
    name: string;
    description?: string | null;
    shortDescription?: string | null;
    price: number;
    salePrice?: number | null;
    sku?: string | null;
    barcode?: string | null;
    stockQuantity?: number;
    isInStock?: boolean;
    isActive?: boolean;
    isFeatured?: boolean;
    isNew?: boolean;
    brandId?: number | null;
    brandName?: string | null;
    categoryId?: number | null;
    categoryName?: string | null;
    slug?: string | null;
    mainImageUrl?: string | null;
    imageUrls?: Array<string> | null;
    tags?: Array<ProductTagDto> | null;
    specifications?: Array<ProductSpecificationDto> | null;
    averageRating?: number;
    ratingCount?: number;
    viewCount?: number;
    createdAt?: string;
    updatedAt?: string | null;
}